Fix PHPCS in Settings_Registry unit tests (XWPENG-51).
Move the WP_Roles test stub into its own file to satisfy OneObjectStructurePerFile and apply PHPCBF alignment fixes.
